faiss_service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
faiss-cpu>=1.9.0
sentence-transformers>=3.3.0
pandas>=2.2.0
pyarrow>=18.0.0
numpy>=2.0.0
pydantic>=2.0.0
